Mocha config (mocha.opts) parse (#53)
What this basically does is allows the user to add a mocha.opts configuration file (https://mochajs.org/#mochaopts) and have it parsed and applied.

The file is not automatically added to the bit component and needs to be specified in the filesRequire array.

We support the following options, which are a subset of the options mocha supports in its programmatic API and the options that are relevant to bit (eg. reporter is not relevant to bit, even though it's supported by the mocha API):
'-b, --bail'
'-g, --grep '
'-s, --slow '
'-t, --timeout '
'--retries '

We also parse the '-r, --require ' option, but since the environment API does not yet support adding a vinyl file through the configuration (I think?) we throw a helpful error, letting the user know they should instead add it to their bit.json file.

The above is also true for including a test file as the last argument of mocha.opts (eg. --slow 12 my-test-file.spec.js) for the same reason.